pyproject.toml
[project]
name = "tryagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
]
